requirements.txt
numpy
sklearn
tensorflow
djangorestframework
